+++ wtk/src/org/apache/pivot/wtk/media/drawing/CubicCurve.java
@@ -180,7 +180,7 @@ public class CubicCurve extends Shape {
         if (!isValid()) {
             // Over-estimate the bounds to keep the logic simple
             int strokeThickness = getStrokeThickness();
-            double radius = ((double)strokeThickness/Math.cos(Math.PI / 4)) / 2;
+            double radius = (strokeThickness / Math.cos(Math.PI / 4)) / 2;
 
             Rectangle2D boundingRectangle = cubicCurve2D.getBounds2D();
             setBounds((int)Math.floor(boundingRectangle.getX() - radius),
