+++ wtk/src/org/apache/pivot/wtk/media/drawing/Line.java
@@ -136,7 +136,7 @@ public class Line extends Shape {
         if (!isValid()) {
             // Over-estimate the bounds to keep the logic simple
             int strokeThickness = getStrokeThickness();
-            double radius = ((double)strokeThickness/Math.cos(Math.PI / 4)) / 2;
+            double radius = (strokeThickness / Math.cos(Math.PI / 4)) / 2;
 
             double top = Math.min(line2D.y1, line2D.y2);
             double left = Math.min(line2D.x1, line2D.x2);
